+++ terraform/hydrocron-dynamo.tf
@@ -105,9 +105,9 @@ resource "aws_dynamodb_table" "hydrocron-reach-track-ingest-table" {
     type = "S"
   }
   global_secondary_index {
-    name               = "statusIndex"
-    hash_key           = "status"
-    projection_type    = "ALL"
+    name            = "statusIndex"
+    hash_key        = "status"
+    projection_type = "ALL"
   }
   point_in_time_recovery {
     enabled = var.stage == "ops" ? true : false
@@ -132,9 +132,9 @@ resource "aws_dynamodb_table" "hydrocron-node-track-ingest-table" {
     type = "S"
   }
   global_secondary_index {
-    name               = "statusIndex"
-    hash_key           = "status"
-    projection_type    = "ALL"
+    name            = "statusIndex"
+    hash_key        = "status"
+    projection_type = "ALL"
   }
   point_in_time_recovery {
     enabled = var.stage == "ops" ? true : false