+++ tests/test_track_ingest.py
@@ -26,7 +26,7 @@ def test_query_cmr():
 
     vcr_cassette = pathlib.Path(os.path.dirname(os.path.realpath(__file__))) \
         .joinpath('vcr_cassettes').joinpath('cmr_query.yaml')
-    with vcr.use_cassette(vcr_cassette):
+    with vcr.use_cassette(vcr_cassette, decode_compressed_response=True):
         track.query_cmr()
         actual_data = track.cmr_granules