+++ CHANGELOG.md
@@ -5,12 +5,22 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+ 
+### Added
+### Changed
+### Deprecated
+### Removed
+### Fixed
+### Security
+
+## [1.5.0]
 
 ### Added
 ### Changed
 ### Deprecated
 ### Removed
 ### Fixed
+    - Issue 243 - Fix cases where not all time steps were returned for a feature id
 ### Security
 
 ## [1.4.0]